refactor backend into directory structure add swaggerui + openapi spec for api testing update gitignore
