+++ docker-compose.yaml
@@ -56,9 +56,16 @@ services:
       - postgres
     volumes:
       - pgadmin_data:/var/lib/pgadmin
+  swagger-ui:
+    image: swaggerapi/swagger-ui:latest
+    ports:
+      - "8081:8080"
+    environment:
+      - SWAGGER_JSON=/openapi/openapi.yaml
+    volumes:
+      - ./backend/openapi.yaml:/openapi/openapi.yaml:ro
 volumes:
   db_data:
   pgadmin_data:
   go-mod-cache:
   go-build-cache:
-
